+++ CHANGES.md
@@ -4,6 +4,10 @@ v0.0.4 (unreleased)
 - command-line tools:
 	- ``waveguide``:
 		- add ability to pass permittivity/permeability of dielectric fill
+- ``rftools/tlines``:
+	- ``Microstrip``:
+		- add method to calculate conductor loss and dielectric loss
+		- add method to calculate attenuation 
 
 v0.0.3 (13-Oct-2020)
 ====================